+++ src/Autoload/Global.gd
@@ -75,6 +75,7 @@ var right_square_indicator_visible := false
 
 # View menu options
 var tile_mode : int = Tile_Mode.NONE
+var mirror_view := false
 var draw_grid := false
 var show_rulers := true
 var show_guides := true