+++ src/Tools/Draw.gd
@@ -257,6 +257,9 @@ func draw_tool_circle(position : Vector2, fill := false) -> void:
 
 
 func draw_tool_brush(position : Vector2) -> void:
+	if Global.mirror_view:
+		position.x = Global.current_project.size.x - position.x
+
 	if Global.tile_mode and _get_tile_mode_rect().has_point(position):
 		position = position.posmodv(Global.current_project.size)
 
@@ -331,6 +334,9 @@ func draw_indicator_at(position : Vector2, offset : Vector2, color : Color) -> v
 
 func _set_pixel(position : Vector2) -> void:
 	var project : Project = Global.current_project
+	if Global.mirror_view:
+		position.x = project.size.x - position.x - 1
+
 	if Global.tile_mode and _get_tile_mode_rect().has_point(position):
 		position = position.posmodv(project.size)