+++ src/UI/Canvas/TileMode.gd
@@ -10,12 +10,21 @@ func _draw() -> void:
 	var positions : Array = get_tile_positions(size)
 	var tilemode_opacity = 1.0 - Global.tilemode_opacity
 
+	var _position := position
+	var _scale := scale
+	if Global.mirror_view:
+		_position.x = _position.x + Global.current_project.size.x
+		_scale.x = -1
+	draw_set_transform(_position, rotation, _scale)
+
 	for i in range(Global.current_project.layers.size()):
 		var modulate_color := Color(1, 1, 1, current_cels[i].opacity - tilemode_opacity)
 		if Global.current_project.layers[i].visible: # if it's visible
 			if Global.tile_mode:
 				for pos in positions:
 					draw_texture(current_cels[i].image_texture, pos, modulate_color)
+
+	draw_set_transform(position, rotation, scale)
 
 
 func get_tile_positions(size):